% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_add_edge.R
\name{map_add_edge}
\alias{map_add_edge}
\title{Add an edge to the mental-health map}
\usage{
map_add_edge(current_matrix, new_edge)
}
\arguments{
\item{current_matrix}{An unweighted directed matrix.}

\item{new_edge}{A vector consisting of two elements, the edge will be included from the first to the second element.}
}
\value{
An updated unweighted directed matrix including the new edge.
}
\description{
Adds a new edge the user has selected to the existing map.
}
\examples{
current_matrix <- matrix(0, nrow = 3, ncol = 3)
colnames(current_matrix) <- rownames(current_matrix) <- c("Stress", "Worrying", "Breakup")
new_edge <- c("Stress", "Breakup")
map_add_edge(current_matrix, new_edge)
}
