% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_create_skeleton.R
\name{map_create_skeleton}
\alias{map_create_skeleton}
\title{Creates a mental-health map skeleton}
\usage{
map_create_skeleton(input_items)
}
\arguments{
\item{input_items}{A vector containing all selected factors.}
}
\value{
A qgraph containing all selected factors and no edges between them.
}
\description{
Creates a mental-health map skeleton
}
\examples{
map_create_skeleton(c("Loss of interest", "Bored", "Self-blame"))
}
