% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_delete_edge.R
\name{map_delete_edge}
\alias{map_delete_edge}
\title{Delete an edge in the mental-health map}
\usage{
map_delete_edge(current_matrix, delete_edge)
}
\arguments{
\item{current_matrix}{An unweighted directed matrix.}

\item{delete_edge}{A vector consisting of two elements, the edge from the first to the second element will be deleted.}
}
\value{
An updated unweighted directed matrix excluding the selected edge.
}
\description{
Deletes the input edge from the existing map.
}
\examples{
current_matrix <- t(matrix(c(0,0,1,0,0,0,0,0,0), nrow = 3, ncol = 3))
colnames(current_matrix) <- rownames(current_matrix) <- c("Stress", "Worrying", "Breakup")
delete_edge <- c("Stress", "Breakup")
map_delete_edge(current_matrix, delete_edge)
}
