% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_delete_node.R
\name{map_delete_node}
\alias{map_delete_node}
\title{Delete a node in the mental-health map}
\usage{
map_delete_node(current_matrix, delete_factor)
}
\arguments{
\item{current_matrix}{An unweighted directed matrix.}

\item{delete_factor}{String input of the factor the user wants to delete.}
}
\value{
An updated unweighted directed matrix excluding the selected factor.
}
\description{
Delete a node the user has selected from their existing map.
}
\examples{
current_matrix <- matrix(0, nrow = 3, ncol = 3)
colnames(current_matrix) <- rownames(current_matrix) <- c("Stress", "Worrying", "Breakup")
delete_factor <- "Stress"
map_delete_node(current_matrix, delete_factor)
}
