% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_find_targets.R
\name{map_find_targets}
\alias{map_find_targets}
\title{Find the most influential factors}
\usage{
map_find_targets(current_matrix, qgraph)
}
\arguments{
\item{current_matrix}{An unweighted directed matrix.}

\item{qgraph}{A qgraph object of current_matrix.}
}
\value{
A string output presenting the most influential factors in the map.
}
\description{
Finds the targets with the highest out-degree centrality (most influential).
}
\examples{
current_matrix <- matrix(0, nrow = 3, ncol = 3)
colnames(current_matrix) <- rownames(current_matrix) <- c("Stress", "Worrying", "Breakup")
qgraph <- qgraph(current_matrix)
map_find_targets(current_matrix, qgraph)
}
