% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_highlight_target.R
\name{map_highlight_target}
\alias{map_highlight_target}
\title{Highlight target factor in the mental-health map}
\usage{
map_highlight_target(current_matrix, target)
}
\arguments{
\item{current_matrix}{An unweighted directed matrix.}

\item{target}{A string denoting the target factor.}
}
\value{
A list node_colors of strings corresponding to the colors of each node in the map.
}
\description{
Highlights the factor the user has selected as a promising treatment target.
}
\examples{
current_matrix <- matrix(0, nrow = 3, ncol = 3)
colnames(current_matrix) <- rownames(current_matrix) <- c("Stress", "Worrying", "Breakup")
target <- "Stress"
map_highlight_target(current_matrix, target)
}
