% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_include_chains.R
\name{map_include_chains}
\alias{map_include_chains}
\title{Include causal chains into mental-health map}
\usage{
map_include_chains(current_matrix, input_chain01, input_chain02)
}
\arguments{
\item{current_matrix}{An unweighted directed matrix.}

\item{input_chain01}{A vector containing all drop down text field inputs of chain 1.}

\item{input_chain02}{A vector containing all drop down text field inputs of chain 2.}
}
\value{
An updated unweighted directed matrix including the two causal chains.
}
\description{
Includes input of chain exercise (PsySys section 2) into mental-health map.
}
\examples{
current_matrix <- matrix(0, nrow = 3, ncol = 3)
colnames(current_matrix) <- rownames(current_matrix) <- c("Stress", "Worrying", "Breakup")
input_chain01 <- c("Stress", "Worrying", "Breakup")
input_chain02 <- c("Worrying", "Stress", "Breakup")
map_include_chains(current_matrix, input_chain01, input_chain02)
}
