% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_include_cycles.R
\name{map_include_cycles}
\alias{map_include_cycles}
\title{Includes input of the chain and vicious cycle exercise into mental-health map}
\usage{
map_include_cycles(
  input_items,
  input_chain01,
  input_chain02,
  input_cycle01,
  input_cycle02
)
}
\arguments{
\item{input_items}{A vector containing all selected factors.}

\item{input_chain01}{A vector containing all drop down text field inputs of chain 1 of the  exercise on PsySys section 2.}

\item{input_chain02}{A vector containing all drop down text field inputs of chain 1 of the  exercise on PsySys section 2.}

\item{input_cycle01}{A vector containing the user input for the first vicious cycle of the exercise on PsySys section 3.}

\item{input_cycle02}{A vector containing the user input for the second vicious cycle of the exercise on PsySys section 3.}
}
\value{
A qgraph containing all selected factors, the causal chains, and the vicious cycles the user selected.
}
\description{
Includes input of the chain and vicious cycle exercise into mental-health map
}
\examples{
map_include_chains(input_items = input$factors,
                   input_chain = c(input$chain01_01, input$chain01_02, input$chain01_03,
                                 input$chain02_01, input$chain02_02, input$chain02_03),
                   input_cycle01 = c(input$cycle1[1], input$cycle1[2]),
                   input_cycle02 = c(input$cycle2[1], input$cycle2[2], input$cycle2[3]))
}
