% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_create_skeleton.R
\name{map_create_skeleton}
\alias{map_create_skeleton}
\title{Create the initial mental-health map}
\usage{
map_create_skeleton(input_items)
}
\arguments{
\item{input_items}{A vector containing all selected factors.}
}
\value{
A 0 matrix with the number of dimensions corresponding to the number of selected factors.
}
\description{
Creates a mental-health map skeleton consisting of the selected factors.
}
\examples{
input_items <- c("Loss of interest", "Bored", "Self-blame")
map_create_skeleton(input_items)
}
