% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_include_chains.R
\name{map_include_chains}
\alias{map_include_chains}
\title{Includes input of chain exercise into mental-health map}
\usage{
map_include_chains(input_items, input_chain01, input_chain02)
}
\arguments{
\item{input_items}{A vector containing all selected factors.}

\item{input_chain01}{A vector containing all drop down text field inputs of chain 1 of the  exercise on PsySys section 2.}

\item{input_chain02}{A vector containing all drop down text field inputs of chain 1 of the  exercise on PsySys section 2.}
}
\value{
A qgraph containing all selected factors and the causal chains the user selected.
}
\description{
Includes input of chain exercise into mental-health map
}
\examples{
map_include_chains(input$factors,
                   c(input$chain01_01, input$chain01_02, input$chain01_03,
                   input$chain02_01, input$chain02_02, input$chain02_03))
}
