% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_include_cycles.R
\name{map_include_cycles}
\alias{map_include_cycles}
\title{Include cycles into mental-health map}
\usage{
map_include_cycles(current_matrix, input_cycle01, input_cycle02)
}
\arguments{
\item{current_matrix}{An unweighted directed matrix.}

\item{input_cycle01}{A vector containing the user input for the first vicious cycle.}

\item{input_cycle02}{A vector containing the user input for the second vicious cycle.}
}
\value{
An updated unweighted directed matrix including the vicious cycles the user selected.
}
\description{
Includes input of the vicious cycle exercise (PsySys section 3) into the mental-health map.
}
\examples{
current_matrix <- matrix(0, nrow = 3, ncol = 3)
colnames(current_matrix) <- rownames(current_matrix) <- c("Stress", "Worrying", "Breakup")
input_cycle01 <- c("Stress", "Worrying")
input_cycle02 <- c("Worrying", "Breakup", "Stress")
map_include_chains(current_matrix, input_cycle01, input_cycle02)
}
